Add parser from POJO to XML

+++ src/main/java/com/xalpol12/entity/Camera.java
@@ -12,8 +12,8 @@ import javax.xml.bind.annotation.XmlType;
 @XmlType(propOrder = {"x", "y", "scaleto", "distance", "node"})
 public class Camera {
     private Node node;
-    private int x;
-    private int y;
+    private double x;
+    private double y;
     private String scaleto;
     private double distance;
 
@@ -23,12 +23,12 @@ public class Camera {
     }
 
     @XmlAttribute
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
     @XmlAttribute
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
